add arduino freertos port and a bunch of other crap :-)

+++ disp/ili9340.h
@@ -21,8 +21,8 @@
 
 #pragma once
 
-#define ILI9340_TFTWIDTH  240
-#define ILI9340_TFTHEIGHT 320
+#define ILI9340_TFTWIDTH  (long)240
+#define ILI9340_TFTHEIGHT (long)320
 
 #define ILI9340_NOP     0x00
 #define ILI9340_SWRESET 0x01
